Abstract out tx
